Move IP-address stuff to `EndpointAddress`.
